backend/src/ClothesSearch/requirements.txt
# Clothes Search Module Requirements
google-genai>=0.2.0
python-dotenv>=1.0.0
fastapi>=0.109.0
pydantic>=2.0.0
